Create TimeTracking
This is an ios shortcut that tracks time spent on a specific activity. "Commute" was put in as a place holder. Could easily change it to "piano practice" or "reading for MaxVel" <- because I should actually be reading for MaxVel right now and figure out pass Hann band filter etc.